recuperacion de funciones en popiedades

+++ classLib/PropiedadDTO.cs
@@ -14,8 +14,8 @@ namespace Entities_DTOs
     public string NombreFinca { get; set; }
     public string Ubicacion { get; set; }
 
-    public decimal Latitud { get; set; }
-    public decimal Longitud { get; set; }
+    public double Latitud { get; set; }
+    public double Longitud { get; set; }
 
     public decimal TamanoHectareas { get; set; }
     public string TipoSuperficie { get; set; }
